sample/iot-smart-grid/requirements.txt
dataclasses-json
hydra-core
azure-storage-file-datalake